# kabtak/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckAccepted": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "poll_after_ms": {
            "default": 2000,
            "title": "Poll After Ms",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "check_id",
          "run_id"
        ],
        "title": "CheckAccepted",
        "type": "object"
      },
...
      "Coverage": {
        "enum": [
          "complete_for_attempted_scope",
          "partial"
        ],
        "title": "Coverage",
        "type": "string"
      },
...
      "EvidenceRead": {
        "properties": {
          "block_id": {
            "title": "Block Id",
            "type": "string"
          },
          "block_sha256": {
            "title": "Block Sha256",
            "type": "string"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "parse_status": {
            "title": "Parse Status",
            "type": "string"
          },
          "parser_version": {
            "title": "Parser Version",
            "type": "string"
          },
          "publisher_role": {
            "title": "Publisher Role",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "version_id": {
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "version_id",
          "block_id",
          "kind",
          "location",
          "text",
          "content_sha256",
          "block_sha256",
          "retrieved_at",
          "parse_status",
          "parser_version",
          "source_url",
          "publisher_role"
        ],
        "title": "EvidenceRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportRead": {
        "additionalProperties": false,
        "properties": {
          "amendments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Amendments",
            "type": "array"
          },
          "application_links": {
            "items": {
              "$ref": "#/components/schemas/ApplicationLinkRead"
            },
            "title": "Application Links",
            "type": "array"
          },
          "conditions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Conditions",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "coverage": {
            "$ref": "#/components/schemas/Coverage"
          },
          "deadline_resolution": {
            "$ref": "#/components/schemas/DeadlineResolution"
          },
          "eligibility": {
            "$ref": "#/components/schemas/EligibilityResolution"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "other_deadlines": {
            "items": {
              "$ref": "#/components/schemas/DeadlineRead"
            },
            "title": "Other Deadlines",
            "type": "array"
          },
          "portal_status": {
            "title": "Portal Status",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/ReportProvenance"
          },
          "reference_time": {
            "format": "date-time",
            "title": "Reference Time",
            "type": "string"
          },
          "required_documents": {
            "items": {
              "$ref": "#/components/schemas/RequiredDocumentRead"
            },
            "title": "Required Documents",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeRead"
          },
          "student_deadline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeadlineRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "mode",
          "reference_time",
          "coverage",
          "scope",
          "deadline_resolution",
          "student_deadline",
          "summary",
          "portal_status",
          "eligibility",
          "provenance"
        ],
        "title": "ReportRead",
        "type": "object"
      },
...
      "RunComparisonRead": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/FactChangeRead"
            },
            "title": "Changes",
            "type": "array"
          },
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "outcome_changed": {
            "title": "Outcome Changed",
            "type": "boolean"
          },
          "previous_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Run Id"
          },
          "profile_changed": {
            "title": "Profile Changed",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "software_versions_changed": {
            "title": "Software Versions Changed",
            "type": "boolean"
          },
          "source_versions_changed": {
            "title": "Source Versions Changed",
            "type": "boolean"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "previous_run_id",
          "classification",
          "summary",
          "outcome_changed",
          "source_versions_changed",
          "profile_changed",
          "software_versions_changed"
        ],
        "title": "RunComparisonRead",
        "type": "object"
      },
      "RunRead": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "reference_time": {
            "format": "date-time",
            "title": "Reference Time",
            "type": "string"
          },
          "report_available": {
            "default": false,
            "title": "Report Available",
            "type": "boolean"
          },
          "retry_of_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of Run Id"
          },
          "stage": {
            "$ref": "#/components/schemas/RunStage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "check_id",
          "kind",
          "status",
          "stage",
          "reference_time",
          "created_at"
        ],
        "title": "RunRead",
        "type": "object"
      },
      "RunStage": {
        "enum": [
          "queued",
          "searching",
          "fetching",
          "extracting",
          "checking",
          "finalizing",
          "finished"
        ],
        "title": "RunStage",
        "type": "string"
      },
      "RunStatus": {
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "interrupted"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deadline and evidence checking API.",
    "title": "Kabtak API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "read_run_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/changes": {
      "get": {
        "operationId": "compare_run_v1_runs__run_id__changes_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunComparisonRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/evidence/{version_id}/{block_id}": {
      "get": {
        "operationId": "read_evidence_v1_runs__run_id__evidence__version_id___block_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "block_id",
            "required": true,
            "schema": {
              "title": "Block Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Evidence",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/report": {
      "get": {
        "operationId": "read_report_v1_runs__run_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Report",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}/retry": {
      "post": {
        "operationId": "retry_run_v1_runs__run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Run",
        "tags": [
          "runs"
        ]
      }
    }
  }
}